Move dataset related classes into a new package
